docs/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras docs/requirements.in
#
alabaster==0.7.16
    # via sphinx
babel==2.14.0
    # via sphinx
beautifulsoup4==4.12.2
    # via furo
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
docutils==0.20.1
    # via sphinx
furo==2023.9.10
    # via -r docs/requirements.in
idna==3.6
    # via requests
imagesize==1.4.1
    # via sphinx
jinja2==3.1.3
    # via sphinx
markupsafe==2.1.3
    # via jinja2
packaging==23.2
    # via sphinx
pyenchant==3.2.2
    # via sphinxcontrib-spelling
pygments==2.17.2
    # via
    #   furo
    #   pygments-github-lexers
    #   sphinx
pygments-github-lexers==0.0.5
    # via -r docs/requirements.in
requests==2.31.0
    # via sphinx
file:.#egg=rpds-py
    # via -r docs/requirements.in
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.5
    # via beautifulsoup4
sphinx==7.2.6
    # via
    #   -r docs/requirements.in
    #   furo
    #   sphinx-basic-ng
    #   sphinx-copybutton
    #   sphinxcontrib-applehelp
    #   sphinxcontrib-devhelp
    #   sphinxcontrib-htmlhelp
    #   sphinxcontrib-qthelp
    #   sphinxcontrib-serializinghtml
    #   sphinxcontrib-spelling
    #   sphinxext-opengraph
sphinx-basic-ng==1.0.0b2
    # via furo
sphinx-copybutton==0.5.2
    # via -r docs/requirements.in
sphinxcontrib-applehelp==1.0.7
    # via sphinx
sphinxcontrib-devhelp==1.0.5
    # via sphinx
sphinxcontrib-htmlhelp==2.0.4
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.6
    # via sphinx
sphinxcontrib-serializinghtml==1.1.9
    # via sphinx
sphinxcontrib-spelling==8.0.0
    # via -r docs/requirements.in
sphinxext-opengraph==0.9.1
    # via -r docs/requirements.in
url-py==0.10.0
    # via -r docs/requirements.in
urllib3==2.1.0
    # via requests
